chore: add mypy to CI, fix 16 type errors across 6 files
- Add requirements-dev.txt (pytest, ruff, mypy, types-PyYAML)
- Update CI to use requirements-dev.txt and add mypy check step
- Add [tool.mypy] config to pyproject.toml (ignore_missing_imports)
- Fix type errors: Promise Literal cast, pandoc float/int, word_counter
  dict type, continuity.py annotation, chapter_validator BannedPattern,
  chapters.py Promise Literal cast

Closes #310

+++ tools/export/pandoc.py
@@ -253,7 +253,7 @@ def assemble_manuscript(
     }
 
 
-def _human_size(size_bytes: int) -> str:
+def _human_size(size_bytes: float) -> str:
     """Convert bytes to human-readable string."""
     for unit in ("B", "KB", "MB", "GB"):
         if size_bytes < 1024:
